[tracing] Generalize {SetTraceValue} method
The {SetTraceValue} method was only defined for a set of integer types,
which sometimes lead to ambiguities when using types like {size_t},
{unsigned long} or the like (see https://crrev.com/c/1886912/1).
This CL fixes that by providing a method accepting any integer type.
It also changes the existing methods to avoid the "cast via union"
idiom, and uses memcpy instead.

R=petermarshall@chromium.org

Bug: v8:9810
Change-Id: I1530405640dc6cb0058153a8dbb860c7f3727ac5