Update patch set 3
Patch Set 3:

I am just wondering. What is a reason behind this change, can this flag be added when feature is ready?

Should protocol return an error when this flag is passed?

Otherwise it will make feature detection much harder in this case. Imagine you are debugging different Node versions from DevTools frontend and would like to check does current node support this flag or not. One of the options might be to go to /json/list endpoint and fetch full protocol definition. This definition would contain this param event if it is not supported.

Patch-set: 3
CC: Gerrit User 1310632 <1310632@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>