Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I am just wondering. What is a reason behind this change, can this flag be added when feature is ready?
> 
> Should protocol return an error when this flag is passed?
> 
> Otherwise it will make feature detection much harder in this case. Imagine you are debugging different Node versions from DevTools frontend and would like to check does current node support this flag or not. One of the options might be to go to /json/list endpoint and fetch full protocol definition. This definition would contain this param event if it is not supported.

Thanks for the comment. We'll introduce protocol changes later in the future.

In this case I don't think it matters. We are rolling out non-spec behavior incrementally, and so there is going to be changes in semantics anyways. And since they are non-spec, nobody should be relying on the behavior except for forwarding it to the users of the repl.

Patch-set: 3