Update patch set 2
Patch Set 2:

So, it looks like my patch won't work for windows. Unfortunately, I don't have a windows machine to test.

According to the manual, vsprintf returns -1 in case of a truncation error (so that explains why my patch fails). And I don't see anything related to conversion or output error.
https://docs.microsoft.com/en-us/cpp/c-runtime-library/reference/vsnprintf-s-vsnprintf-s-vsnprintf-s-l-vsnwprintf-s-vsnwprintf-s-l?view=vs-2019


Should I assume that a conversion error will not happen on Windows? If so, what about continuing to return -1 in case of truncation, and returning -2, in case of a conversion error? (that return value would only be returned on posix platforms).

Patch-set: 2