Update patch set 5
Patch Set 5:

Thanks a lot for tackling this, it has been on my to-do list for quite a while 😊

I think this CL can be greatly simplified: Whenever we create a template specialization, we create a Scope to declare the generic parameters in. For generic callables, this scope is the callable declarable itself. For generic types, it's a namespace created just for this purpose. While generating code for the template specialization, we always have access to this scope since we use it to lookup the template parameters. So by walking the scope chain, we'll always reach this scope, even for say struct methods. So this CL would simplify to:
1. Put the SpecializationRequester into this scope at it's creation site, of which there are only 2: DeclarationVisitor::Specialize() and TypeOracle::CreateGenericTypeInstatiationNamespace()
2. Walk the scope chain when generating an error to find the respective scope.

What do you think?

Patch-set: 5
Reviewer: Gerrit User 1180419 <1180419@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>