Update patch set 2
Patch Set 2:

It looks like we never staged this previously, and I'm a little reluctant to turn this flag all the way on so close to the branch point without having any Clusterfuzz coverage.

Is there a reason to think this is particularly low-risk?

Patch-set: 2
Reviewer: Gerrit User 1115851 <1115851@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>