Update patch set 5
Patch Set 5:

Isn't the mismatch actually in CodeStubAssembler::ComputeSeededHash, where we truncate the key to Int32, but specify the parameter to be of type Uint32?
https://cs.chromium.org/chromium/src/v8/src/codegen/code-stub-assembler.cc?sq=package:chromium&g=0&l=7988

So with this change, we should also change the specified parameter type to Int32.

Patch-set: 5
Reviewer: Gerrit User 1177599 <1177599@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>