Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> The approach lgtm, but I know nothing about Mips assembly. Could someone who knows Mips review this?

lgtm, thank you.

this patch ensure do not change the input value.

maybe we can do some performance improvement in the future: 
we do not need sign extend i32 to i64 (sll r, r, 0)all the time, we only need do it if the follow instruction is a branch instruction which will consider the input value as 64bit, if the follow instruction is a 'addi' or something else we do not need do this, this instruction will consider the input value as a 32bit value and the hardware ensure the result is sign extended to 64 bit.

Patch-set: 5
Reviewer: Gerrit User 1317343 <1317343@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>