Make some tests thread safe
This CL changes some of the tests to be thread
safe by adding the postfix "DeathTest" to their
names; because these tests failed when ran in
parallel and adding the DeathTest postfix makes
them run before all other tests and subsequently
making them thread safe. A few functions had to
be made available to other classes using the
friend declaration following this change.

Bug: chromium:934932
Change-Id: I9578a4d1a50607522aba3111439f95b1037e777f