Update patch set 6
Patch Set 6:

(1 comment)

If I read the CL description correctly, we only need to add the "DeathTest" suffix to force the tests to run sequentially. It's not (always) the case that it's actual death tests, right?
In that case we should find another way to force these tests to run individually on Fuschia, without "misusing" the "DeathTest" feature. Can we discuss this with more people (especially machenbach) next week?

Patch-set: 6
CC: Gerrit User 1177599 <1177599@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>