Update patch set 7
Patch Set 7:

CL description says that this is about thread-safety, but then says that we need to run tests sequentially rather than in-parallel.  Parallel tests are run in separate batch processes, so thread-safety isn't an issue here, I don't believe?

What is the actual failure-mode that this CL is intended to address?

Patch-set: 7
CC: Gerrit User 1001385 <1001385@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>