[gcmole] Ignore calls guarded by DisallowHeapAllocation
Although the per-function analysis checked for DisallowHeapAllocation
scopes, the callee-based gcsuspects creation didn't, so a function calling
a gcsuspect inside a DisallowHeapAllocation scope would get marked as a
gcsuspect. So, for the case

    void foo() {
        DisallowHeapAllocation no_gc;
        maybe_call_gc();
    }

    void bar() {
        Object obj;
        foo();
        obj.address();
    }

foo would be marked as a gcsuspect, and so obj.address() would be
considered a dead access to obj, despite us explicitly disallowing
allocation.

Now, CalleesPrinter explicitly checks for GC guards, and skips functions
called while guarded.

As a drive-by, skip outputting functions which don't call anything. They
will still be resolved if they are called, so we don't need to waste time
printing and parsing them at their definition site.

Bug: v8:9680
Change-Id: Id05f839327c4947afa0c55cf35faffb561e51201