Update patch set 2
Patch Set 2: Code-Review+1

LGTM, thanks!

My 2 cents on the IIFE is that it's a widely used pattern in JS in general, so it's a good practice to use it our tests too. Leaving out the +2, so that Michael can see it as well.

Patch-set: 2
Reviewer: Gerrit User 1283477 <1283477@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1