Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> RSLGTM. Can we also fix it on other architectures for consistency?

ARM64 is the only other platform using pointer compression currently. I added the change for that, but can't test locally. (It's a very simple mechanical change to one line however).

I added a test that does crash before this CL also, so should be ready to go now if looks good to folks.

Patch-set: 2