[cpu-profiler] Add a flag to run the CPU profiler in D8
This flag enables the CPU profiler before running any scripts in D8.

This is so that we can turn the flag on during fuzzing to help us find
security and stability issues.

Right now this flag has problems with a bunch of tests under
mjsunit/asm which fail in the stack walker when the flag is on. I'll
address these in a follow-up before we turn this on for fuzzing.

Bug: v8:10150

Change-Id: Ief8134ab2e9da63851b9c235f47ba217d1dce348