Update patch set 3
Patch Set 3: Commit-Queue+2

> Patch Set 3: Code-Review+1
> 
> lgtm (there seems to be only a single production caller of this method, but I guess that means it'll be easy to deprecate 😊)

Ack, but we still need to do this deprecation dance :/
The plan is to do this till the end of this week, so the deprecation is in the M-81 branch. We can then remove the old method right after the branch.

Thanks for review, landing.

Patch-set: 3
Reviewer: Gerrit User 1177599 <1177599@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+2