Update patch set 1
Patch Set 1: Code-Review+1

Thanks for prototyping this, and putting in the work to find the optimal code sequence. 

Context for Tobias: Zhi prototyped, and implemented this to form an ergonomic set of SIMD 64x2 operations, we brought this up to the community because even the best variant of this on 32-bit platforms is not great, and it was decided that these operations will not be a part of the specification.

Patch-set: 1
Reviewer: Gerrit User 1122079 <1122079@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1