[api] Add IsContextValidInHost callback
To schedule FinalizationGroup callbacks from within V8, we need to be
able to tell if a context is still considered valid by the host. In
blink, this means either the v8::Context has a nullptr embedder data in
kV8ContextPerContextDataIndex, or !ScriptState::From(v8_context)->
ContextIsValid(), e.g. if the frame related to the context has navigated
or been removed.

If the host doesn't install a callback, all contexts are always
considered valid.

Bug: v8:8179
Change-Id: I8c438f710630024d16a3d1153022c7727bff5568