[interpreter][builtins] Make FunctionEntry StackChecks implicit
FunctionEntry StackChecks is one of the two cases where we generate a
StackCheck bytecode. In these cases, we do stack check against the js limit
(not to be confused with the real js limit). Their purpose is to be able
to interrupt the running code.

We can omit the FunctionEntry StackCheck by embedding its code into
the InterpreterEntryTrampoline builtin. We save one bytecode per
interpreted function

Bug: v8:10149, v8:9977
Change-Id: I6156de48b3bc0b519dd21190a8e6214fbe96c78d