heap: Delay completing marking
Delay completing marking (and thus the atomic GC pause) during JS executions,
increasing the chance to finalize the garbage collection from a task. This is
beneficial as it avoids stack scanning which is expensive and can keep alive
outdated objects in case of unified heap.

The delay is limited by
- 10% of the incremental marking walltime
- weighted average of previously recorded time to incremental marking task
  invocations

Bug: chromium:1044630
Change-Id: I10e63e6aaa88d8488d4415f311016dce2b4e62a2