[wasm-simd] Force some shuffles to use register for src0 and src1
test-run-wasm-simd/RunWasm_S8x16MultiShuffleFuzz_turbofan was failing
reliably with --no-enable-avx. (Even though the shuffle sequences were
randomly generated, in practice we quite quickly hit a case where we
will get a segfault.)

For 32x4swizzle and 32x4 shuffle, they use pshufd, which can take an
operand, but needs to be 16-byte aligned, which they are not, current.
So force them to be registers for now. This is similar to what we do in
the x64 selection too.

Bug: v8:9198
Change-Id: If319ff276202d4be095714a6cb18dec0d0551efd