[wasm] Custom sections should not cause validation errors
The {WasmSectionIterator} used the normal {Decoder} object to decode
the name of a custom section. However, if there is a problem in the
name, then this would cause a validation error. With this change we use
a separate {Decoder} object to decode the name. Thereby an error in the
name decoding will not cause a validation error.

R=thibaud@chromium.org

Fix: v8:10126
Change-Id: I672393df4f39405109dff7cb49e53bf981169e07