Disable oneshot optimizations
We used to optimize functions that are expected to executed only
once by not allocating feedback slots for some of the bytecodes. This
would help in reducing the memory and avoiding initializing feedback
that would be never used. With lazy feedback allocation, we don't
allocate feedback vectors for most of such functions anyway.

The generated bytecode for oneshot optimized functions is different and
if we don't properly track this information we might end up generating
different bytecode for the same function. This could causes problems
when there is a mismatch between the feedback slots used by the new
bytecode and the old bytecode.

Since we potentially get most of the benefits of one shot optimization
with lazy feedback vector allocation we can simplify the code by
disabling this optimization.

Bug:chromium:1045824
Change-Id: Ib94605c8c766adc99f54c8333f780d2448caff5e