Update patch set 5
Patch Set 5:

(7 comments)

Being that I've reviewed a bunch of this code previously, (and for a small portion - wrote it), I've given it a quick skim over and left a couple notes. I'll wait to see if others have any "big picture" concerns before I revisit in detail. Thanks!

Patch-set: 5
CC: Gerrit User 1322303 <1322303@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>